PSEC-2031 move pytest fixtures into conftest.py
